+++ examples/load_balancer/lb_full/lb_full.tf
@@ -314,6 +314,20 @@ resource "oci_load_balancer_listener" "lb-listener2" {
   }
 }
 
+resource "oci_load_balancer_listener" "lb-listener3" {
+  load_balancer_id         = "${oci_load_balancer.lb1.id}"
+  name                     = "tcp"
+  default_backend_set_name = "${oci_load_balancer_backend_set.lb-bes1.name}"
+  port                     = 80
+  protocol                 = "TCP"
+  rule_set_names           = ["${oci_load_balancer_rule_set.test_rule_set.name}"]
+
+  connection_configuration {
+    idle_timeout_in_seconds            = "2"
+    backend_tcp_proxy_protocol_version = "1"
+  }
+}
+
 resource "oci_load_balancer_backend" "lb-be1" {
   load_balancer_id = "${oci_load_balancer.lb1.id}"
   backendset_name  = "${oci_load_balancer_backend_set.lb-bes1.name}"